Methods(return)
In this simple task I was had create a conditions into the signIn method. It'll identify name and surname if the user. If the name is "user" program will stop. If not it will continue, but the program won't give an output. Propose of the task is to make clear about return statement.